Update nav menue & add css to multiple pages
