+++ package.json
@@ -37,7 +37,8 @@
     "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.1.3",
-    "styled-components": "^5.3.6"
+    "styled-components": "^5.3.6",
+    "sweetalert2": "^11.6.14"
   },
   "scripts": {
     "start": "node server/server.js",
