+++ src/components/DailyEntry/DailyEntry.css
@@ -2,7 +2,7 @@
     justify-content: center;
     align-items: center;
     margin: 0;
-    padding-left: 20px;
+    padding: 20px;
     font-family: 'Alice', serif;
 }
