+++ src/components/EntryHistory/EntryHistory.jsx
@@ -1,14 +1,161 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector} from 'react-redux';
-// import { useHistory } from 'react-router-dom';
+import { useHistory, useParams } from 'react-router';
+
+// MUI
+import PropTypes from 'prop-types';
+import Box from '@mui/material/Box';
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell from '@mui/material/TableCell';
+import TableContainer from '@mui/material/TableContainer';
+import TableHead from '@mui/material/TableHead';
+import TablePagination from '@mui/material/TablePagination';
+import TableRow from '@mui/material/TableRow';
+import TableSortLabel from '@mui/material/TableSortLabel';
+import Paper from '@mui/material/Paper';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import Switch from '@mui/material/Switch';
+import IconButton from '@mui/material/IconButton';
+import DeleteIcon from '@mui/icons-material/Delete';
+import EditTwoToneIcon from '@mui/icons-material/EditTwoTone';
+import { visuallyHidden } from '@mui/utils';
+
+function descendingComparator(a, b, orderBy) {
+  if (b[orderBy] < a[orderBy]) {
+    return -1;
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1;
+  }
+  return 0;
+}
+
+function getComparator(order, orderBy) {
+  return order === 'desc'
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy);
+}
+
+// This method is created for cross-browser compatibility, if you don't
+// need to support IE11, you can use Array.prototype.sort() directly
+function stableSort(array, comparator) {
+  const stabilizedThis = array.map((el, index) => [el, index]);
+  stabilizedThis.sort((a, b) => {
+    const order = comparator(a[0], b[0]);
+    if (order !== 0) {
+      return order;
+    }
+    return a[1] - b[1];
+  });
+  return stabilizedThis.map((el) => el[0]);
+}
+
+const headCells = [
+  {
+    id: 'prescription_name',
+    numeric: false,
+    disablePadding: true,
+    label: 'Prescription Name ',
+  },
+  {
+    id: 'prescription_amount',
+    numeric: true,
+    disablePadding: false,
+    label: 'Dosage',
+  },
+  {
+    id: 'addDate',
+    numeric: true,
+    disablePadding: false,
+    label: 'Time Taken ',
+  },
+  {
+    id: 'quantity',
+    numeric: true,
+    disablePadding: false,
+    label: 'Quantity Taken ',
+  },
+  {
+    id: 'notes ',
+    numeric: true,
+    disablePadding: false,
+    label: 'Notes ',
+  },
+  {
+    id: 'edit',
+    numeric: true,
+    disablePadding: false,
+    label: 'Edit',
+  },
+  {
+    id: 'delete',
+    numeric: true,
+    disablePadding: false,
+    label: 'Delete',
+  },
+];
+
+function EnhancedTableHead(props) {
+  const { order, orderBy, onRequestSort } =
+    props;
+  const createSortHandler = (property) => (event) => {
+    onRequestSort(event, property);
+  };
+
+  return (
+    <TableHead sx={{ bgcolor: '#a6c7f4' }} >
+      <TableRow>
+        <TableCell >
+        </TableCell>
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id}
+            align={headCell.numeric ? 'right' : 'left'}
+            padding={headCell.disablePadding ? 'none' : 'normal'}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            <TableSortLabel
+              active={orderBy === headCell.id}
+              direction={orderBy === headCell.id ? order : 'asc'}
+              onClick={createSortHandler(headCell.id)}
+            >
+              {headCell.label}
+              {orderBy === headCell.id ? (
+                <Box component="span" sx={visuallyHidden}>
+                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
+                </Box>
+              ) : null}
+            </TableSortLabel>
+          </TableCell>
+        ))}
+        <TableCell >
+        </TableCell>
+      </TableRow>
+    </TableHead>
+  );
+}
+
+EnhancedTableHead.propTypes = {
+  onRequestSort: PropTypes.func.isRequired,
+  order: PropTypes.oneOf(['asc', 'desc']).isRequired,
+  orderBy: PropTypes.string.isRequired,
+  rowCount: PropTypes.number.isRequired,
+};
 
 function EntryHistory() {
-    console.log('in EntryHistory');
+  const [order, setOrder] = React.useState('asc');
+  const [orderBy, setOrderBy] = React.useState('dosage');
+  const [page, setPage] = React.useState(0);
+  const [dense, setDense] = React.useState(false);
+  const [rowsPerPage, setRowsPerPage] = React.useState(5);
+
+  console.log('in EntryHistory');
 
     const dispatch = useDispatch();
-    // const history = useHistory();
-    const entryHistory = useSelector(store => store.dailyEntryReducer);
-    console.log('user history ****', entryHistory);
+    const history = useHistory();
+    const rows = useSelector(store => store.dailyEntryReducer);
+    console.log('user history ****', rows);
 
     useEffect(() => {
         dispatch({
@@ -16,55 +163,119 @@ function EntryHistory() {
         })
     }, [])
 
-    return (
-        <>
-            <h3>Entry History Page</h3>
-            
-                <table>
-                    <thead>
-                        <tr>
-                            <th>Prescription Name</th>
-                            <th>Prescription Amount</th>
-                            <th>Time Taken</th>
-                            <th>Quantity</th>
-                            <th>Notes</th>
-                            <th>Edit Entry</th>
-                        </tr>
-                        
-                        {entryHistory.map(entry => {
-                            return (
-                                <tr key={entry.id}>
-                                    <td>
-                                        {entry.prescription_name}
-                                    </td>
+  const handleRequestSort = (event, property) => {
+    const isAsc = orderBy === property && order === 'asc';
+    setOrder(isAsc ? 'desc' : 'asc');
+    setOrderBy(property);
+  };
 
-                                    <td>
-                                        {entry.prescription_amount}
-                                    </td>
+  const handleChangePage = (event, newPage) => {
+    setPage(newPage);
+  };
 
-                                    <td>
-                                        {entry.addDate}
-                                    </td>
+  const handleChangeRowsPerPage = (event) => {
+    setRowsPerPage(parseInt(event.target.value, 10));
+    setPage(0);
+  };
 
-                                    <td>
-                                        {entry.quantity}
-                                    </td>
+  const handleChangeDense = (event) => {
+    setDense(event.target.checked);
+  };
 
-                                    <td>
-                                        {entry.notes}
-                                    </td>
+  // Avoid a layout jump when reaching the last page with empty rows.
+  const emptyRows =
+    page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
 
-                                    <td>
-                                        <button>Edit Entry</button>
-                                    </td>
+  return (
+    <Box sx={{ width: '100%' }}>
+      <Paper sx={{ width: '100%', mb: 2 }}>
+        <TableContainer>
+          <Table
+            sx={{ minWidth: 750 }}
+            aria-labelledby="tableTitle"
+            size={dense ? 'small' : 'medium'}
+          >
+            <EnhancedTableHead
+              order={order}
+              orderBy={orderBy}
+              onRequestSort={handleRequestSort}
+              rowCount={rows.length}
+            />
+            <TableBody>
+              {/* if you don't need to support IE11, you can replace the `stableSort` call with:
+                 rows.sort(getComparator(order, orderBy)).slice() */}
+              {stableSort(rows, getComparator(order, orderBy))
+                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+                .map((row, index) => {
 
-                                </tr>
-                            )
-                        })}
-                    </thead>
-                </table>
-        </>
-    );
+                  return (
+                    <TableRow
+                      hover
+                      key={row.id} // added id here
+                    >
+                      <TableCell padding="checkbox">
+                      </TableCell>
+                      <TableCell
+                        component="th"
+                        scope="row"
+                        padding="none"
+                      >
+                        {row.prescription_name}
+                      </TableCell>
+                      <TableCell align="right">{row.prescription_amount}</TableCell>
+                      <TableCell align="right">{new Date(row.addDate).toLocaleString()}</TableCell>
+                      <TableCell align="right">{row.quantity}</TableCell>
+                      <TableCell align="right">{row.notes}</TableCell>
+                      <TableCell align="right"> 
+                          <IconButton 
+                            label="Edit Entry"
+                            variant="outlined"
+                            onClick={() => dispatch({ type: 'DELETE_ENTRY', payload: row.id })}
+                          > 
+                            <EditTwoToneIcon /> 
+                          </IconButton>
+                        </TableCell>
+                      <TableCell align="right"> 
+                          <IconButton 
+                            aria-label="delete"
+                            onClick={() => dispatch({ type: 'DELETE_ENTRY', payload: row.id })}
+                          > 
+                            <DeleteIcon /> 
+                          </IconButton>
+                        </TableCell>
+                        <TableCell padding="checkbox">
+                      </TableCell>
+                    </TableRow>
+                  );
+                })}
+              {emptyRows > 0 && (
+                <TableRow
+                  style={{
+                    height: (dense ? 33 : 53) * emptyRows,
+                  }}
+                >
+                  <TableCell colSpan={6} />
+                </TableRow>
+              )}
+            </TableBody>
+          </Table>
+        </TableContainer>
+        <TablePagination
+          rowsPerPageOptions={[5, 10, 25]}
+          component="div"
+          count={rows.length}
+          rowsPerPage={rowsPerPage}
+          page={page}
+          onPageChange={handleChangePage}
+          onRowsPerPageChange={handleChangeRowsPerPage}
+        />
+      </Paper>
+      <FormControlLabel
+        control={<Switch checked={dense} onChange={handleChangeDense} />}
+        label="Dense padding"
+      />
+    </Box>
+  );
 }
 
 export default EntryHistory;
