+++ src/components/Nav/Nav.css
@@ -17,7 +17,7 @@
   font-size: 24px;
   font-weight: 700;
   color: #f2f2f2;
-  padding-left: 10px;
+  padding: 10px;
   margin: 0;
 }
