+++ src/components/Nav/Nav.jsx
@@ -3,23 +3,17 @@ import { Link } from 'react-router-dom';
 import LogOutButton from '../LogOutButton/LogOutButton';
 import './Nav.css';
 import { useSelector } from 'react-redux';
-
+import NavDrawer from '../NavDrawer/NavDrawer';
 function Nav() {
   const user = useSelector((store) => store.user);
 
   return (
     <div className="nav">
+      <NavDrawer />
       <Link to="/home">
         <h2 className="nav-title">TrackPreScripts</h2>
       </Link>
       <div>
-        {/* If no user is logged in, show these links */}
-        {!user.id && (
-          // If there's no user, show login/registration links
-          <Link className="navLink" to="/login">
-            Login / Register
-          </Link>
-        )}
 
         {/* If a user is logged in, show these links */}
         {user.id && (
@@ -52,9 +46,17 @@ function Nav() {
           </>
         )}
 
-        {/* <Link className="navLink" to="/about">
+        {/* If no user is logged in, show these links */}
+        {!user.id && (
+          // If there's no user, show login/registration links
+          <Link className="navLink" to="/login">
+            Login / Register
+          </Link>
+        )}
+
+        <Link className="navLink" to="/about">
           About
-        </Link> */}
+        </Link>
       </div>
     </div>
   );
