+++ src/components/PrescriptionForm/PrescriptionForm.jsx
@@ -2,13 +2,13 @@ import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom'
 import React, { useState } from 'react';
 
+import PrescriptionSelect from './PrescriptionSelect';
+
 // Material-UI imports listed down below
 import Button from '@mui/material/Button';
-import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
-import Typography from '@mui/material/Typography';
-import FormControl from '@mui/material/FormControl';
 import Chip from '@mui/material/Chip';
+import Stack from '@mui/material/Stack';
 
 function PrescriptionForm() {
     console.log('in PrescriptionForm');
@@ -29,89 +29,98 @@ function PrescriptionForm() {
             payload: newPrescription
         })
         console.log('here is a newPrescription', newPrescription);
-        // history.push('/home');
+        history.push('/home');
     }
 
     return (
         <>
-            <Typography variant='h3' mt={2} sx={{ p: 3 }}>
-                Add New Prescription Form
-            </Typography>     
+            <div className='title'>
+                    Add New Prescription
+            </div> 
 
-            <Box 
+            <Stack
                 className='myBox'
                 component="form" 
                 mx={'auto'}
                 sx={{ 
-                    '& .MuiTextField-root': { p: 6, width: '18rem', height: '4rem' ,bgcolor: 'white',  },
+                    '& .MuiTextField-root': { p: 6, width: '18rem', height: '4rem' ,bgcolor: 'white', borderRadius: 5 },
                     color: 'text.primary'
                 }}
                 noValidate
                 autoComplete="off"
+                spacing={1}
                 onSubmit={(event) => addPrescription(event)}
             >
-                <FormControl
-                    sx= {{ bgcolor: 'grey.400'}}
-                >
-                <TextField
-                    required
-                    id="filled-required"
-                    label="Prescription Name"
-                    placeholder="Prescription Name"
-                    variant="filled"
-                    onChange={onHandlePrescription}
-                    name="prescription"
-                />
-
-                <TextField
-                    required
-                    id="filled-required"
-                    label="Dosage"
-                    placeholder="Dosage"
-                    variant="filled"
-                    onChange={onHandlePrescription}
-                    name="dosage"
-                />
-
-                <TextField
-                    required
-                    id="filled-required"
-                    label="Count"
-                    placeholder="Count"
-                    variant="filled"
-                    onChange={onHandlePrescription}
-                    name="count"
-                />
-
-                <TextField
-                    required
-                    id="outlined-multiline-flexible"
-                    label="Description"
-                    placeholder="Description"
-                    variant="filled"
-                    multiline
-                    maxRows={4}
-                    onChange={onHandlePrescription}
-                    name="description"
-                />
-
-                <Button
-                    type="submit"
-                    variant="contained"
-                    sx={{ p: 2 }}
-                >
-                    Save Entry
-                </Button>
-                </FormControl>
                 
-            </Box>
+            <TextField
+                required
+                id="filled-required"
+                label="Prescription Name"
+                placeholder="Prescription Name"
+                variant="filled"
+                onChange={onHandlePrescription}
+                name="prescription"
+            />
+
+            {/* <PrescriptionSelect 
+                required
+                id="filled-required"
+                label="Prescription Name"
+                placeholder="Prescription Name"
+                variant="filled"
+                onChange={onHandlePrescription}
+                name="prescription"
+            /> */}
+
+            <TextField
+                required
+                id="filled-required"
+                label="Dosage"
+                placeholder="Dosage"
+                variant="filled"
+                onChange={onHandlePrescription}
+                name="dosage"
+            />
+
+            <TextField
+                required
+                id="filled-required"
+                label="Count"
+                placeholder="Count"
+                variant="filled"
+                onChange={onHandlePrescription}
+                name="count"
+            />
+
+            <TextField
+                required
+                id="outlined-multiline-flexible"
+                label="Description"
+                placeholder="Description"
+                variant="filled"
+                multiline
+                maxRows={4}
+                onChange={onHandlePrescription}
+                name="description"
+            />
+
+            <Button
+                type="submit"
+                variant="contained"
+                sx={{ p: 2 }}
+            >
+                Save Entry
+            </Button>
             
             <Chip 
-                label="Back to My MedList"
+                label="Cancel"
                 color="primary"
                 sx={{ m: 4, height: 40 }}
                 onClick={() => { history.push('/home')}}
             />
+            </Stack>
+
+            
         </>
     );
 }
